Add support for loading custom vocabulary configs in realtime
-See code validator for implementation
-2 new properties added in referenceccdaservice.xml
--referenceccda.isDynamicVocab
---Set to false if wating to use a single pre-loaded default config only
like past versions of the validator
---Set to true for the option to use one or more custom dynamically
loaded configs determined by API form-data vocabularyConfig key which
should have a value set to the name of the file excepting the .xml
extension
--referenceccda.configFolder
---Set to the path to hold the configs. This can be the same directory
as the current default config path or a completely different directory.
Error handling is implemented to handle either scenario. If for any
reason a custom dynamic config is not found, the default config will be
loaded, if it exists.