Return total validations run based on count in configured XML
-The Code Validator API now provides a getter to the new
"GlobalCodeValidatorResults" which currently only contains a
"vocabularyValidationConfigurationsCount." We are now returning
vocabularyValidationConfigurationsCount in the JSON which is
semi-dynamic in that it tracks the total vocabulary validations
configured for that particular use (validation done) by the code
validator. Since the configuration can be dynamicaly swapped for each
instance, a different count can be tracked in this case. Otherwise, if
the configuration is static, the count would remain static for the
duration of the referenceccdaservice deployment.
--example of JSON in referenceccdaservice
---"vocabularyValidationConfigurationsCount":179
-see
https://github.com/siteadmin/code-validator-api/commit/5c8edf43c6b89e5b71de113807affd448f62f6b7